chore(api): move webhook tests out of generated code
